--- src/main/java/com/wine/to/up/simple/parser/service/SimpleParser/ShopCatalog.java
@@ -1,28 +1,0 @@
-package com.wine.to.up.simple.parser.service.SimpleParser;
-
-import lombok.*;
-
-import java.io.FileWriter;
-import java.io.IOException;
-
-@Data
-@AllArgsConstructor
-public class ShopCatalog {
-
-    private int wineID;
-    private int wineShopID;
-    private int shopID;
-    private int price;
-    private WineShop newShop;
-
-    public void writeToFile(String filePath) {
-        try (FileWriter fileWriter = new FileWriter(filePath, true)) {
-            fileWriter.write(this.toString());
-
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-
-    }
-}
-
